Re-expose modal and confirm components
