+++ packages/components/src/Modal/Modal.tsx
@@ -2,38 +2,51 @@ import * as React from "react"
 import styled from "react-emotion"
 import { OperationalStyleConstants } from "../utils/constants"
 
-import Overlay from "../Internals/Overlay"
-import Card, { Props as CardProps } from "../Card/Card"
+import Button, { Props as ButtonProps } from "../Button/Button"
+import ControlledModal from "../Internals/ControlledModal"
 
-export interface Props {
-  id?: string
-  className?: string
-  /** Content class name */
-  contentClassName?: string
-  /** Children */
-  children: CardProps["children"]
-  onClose?: () => void
-  /** Title */
-  title?: CardProps["title"]
-  /** Action(s) to appear in top-right */
-  action?: CardProps["action"]
+export interface ModalOptions {
+  title: React.ReactNode
+  body: React.ReactNode | ((close: () => void) => React.ReactNode)
 }
 
-const Content = styled(Card)(({ theme }: { theme?: OperationalStyleConstants }) => ({
-  position: "fixed",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  zIndex: theme.zIndex.modal,
-}))
+export interface State {
+  options?: ModalOptions
+}
 
-const Modal: React.SFC<Props> = (props: Props) => (
-  <>
-    <Overlay id={props.id} className={props.className} onClick={props.onClose} />
-    <Content className={props.contentClassName} title={props.title} action={props.action}>
-      {props.children}
-    </Content>
-  </>
-)
+export interface Props {
+  children: (modal: (options: ModalOptions) => void) => React.ReactNode
+}
+
+export class Modal extends React.Component<Props, Readonly<State>> {
+  readonly state: State = {
+    options: undefined,
+  }
+
+  openModal = (options: ModalOptions) => {
+    this.setState({ options })
+  }
+
+  closeModal = () => {
+    this.setState({ options: undefined })
+  }
+
+  render() {
+    const isOpen = Boolean(this.state.options)
+
+    return (
+      <>
+        {typeof this.props.children === "function" ? this.props.children(this.openModal) : this.props.children}
+        {isOpen && (
+          <ControlledModal title={this.state.options.title} onClose={this.closeModal}>
+            {typeof this.state.options.body === "function"
+              ? this.state.options.body(this.closeModal)
+              : this.state.options.body}
+          </ControlledModal>
+        )}
+      </>
+    )
+  }
+}
 
 export default Modal
