+++ packages/components/src/index.ts
@@ -16,6 +16,7 @@ export { default as CardColumn } from "./CardColumn/CardColumn"
 export { default as CardColumns } from "./CardColumns/CardColumns"
 export { default as Chip } from "./Chip/Chip"
 export { default as Code } from "./Code/Code"
+export { default as Confirm } from "./Confirm/Confirm"
 export { default as ContextMenu } from "./ContextMenu/ContextMenu"
 export { default as Logo } from "./Logo/Logo"
 export { default as Grid } from "./Grid/Grid"
